+++ packages/core/acme-client/src/axios.js
@@ -18,7 +18,8 @@ instance.defaults.headers.common['User-Agent'] = `node-${pkg.name}/${pkg.version
 /* Default ACME settings */
 instance.defaults.acmeSettings = {
     httpChallengePort: 80,
-    httpsChallengePort: 443
+    httpsChallengePort: 443,
+    tlsAlpnChallengePort: 443
 };